refactor: Rename and imports to BalanceInput component
