+++ src/components/BalanceInput/index.ts
@@ -1,0 +1,2 @@
+export { default } from './BalanceInput'
+export type { BalanceInputProps } from './BalanceInput'
